moved anikifor manual tests
