+++ cfme/tests/cloud/test_providers.py
@@ -251,50 +251,6 @@ def test_region_required_validation(request, soft_assert, appliance):
         soft_assert(view.region.help_block == "Required")
 
 
-@pytest.mark.manual
-@pytest.mark.tier(1)
-def test_regions_all_azure():
-    """
-    Need to validate the list of regions we show in the UI compared with
-    regions.rb  Recent additions include UK South
-    These really don"t change much, but you can use this test case id
-    inside bugzilla to set qe_test flag.
-
-    Polarion:
-        assignee: anikifor
-        casecomponent: Cloud
-        caseimportance: medium
-        initialEstimate: 1/12h
-        startsin: 5.6
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(1)
-def test_regions_disable_azure():
-    """
-    https://bugzilla.redhat.com/show_bug.cgi?id=1412355
-    CloudForms should be able to enable/disable unusable regions in Azure,
-    like the Government one for example.
-
-    Polarion:
-        assignee: anikifor
-        casecomponent: Cloud
-        caseimportance: medium
-        initialEstimate: 1/10h
-        testSteps:
-            1. Go into advanced settings and add or remove items from the following section.
-               :ems_azure:
-               :disabled_regions:
-               - usgovarizona
-               - usgoviowa
-               - usgovtexas
-               - usgovvirginia
-    """
-    pass
-
-
 @pytest.mark.tier(3)
 @test_requirements.discovery
 def test_host_name_required_validation_cloud(request, appliance):
@@ -767,90 +723,6 @@ def test_domain_id_validation(request, provider):
 
     # ToDo: Assert proper flash message after BZ-1545520 fix.
     assert view.flash.messages[0].type == 'error'
-
-
-@pytest.mark.manual
-@pytest.mark.tier(1)
-def test_public_ip_without_nic_azure():
-    """
-    Relates to https://bugzilla.redhat.com/show_bug.cgi?id=1531099  -
-    Update testcase after BZ gets resolved
-    Update: we are not filtering PIPs but we can use PIPs which are not
-    associated to any NIC
-
-    Polarion:
-        assignee: anikifor
-        casecomponent: Cloud
-        caseimportance: medium
-        initialEstimate: 1/6h
-        testSteps:
-            1. Have a Puplic IP on Azure which is not assigned to any Network
-            Interface - such Public IPs should be reused property
-            2. Provision Azure Instance - select public IP from 1.
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(1)
-def test_sdn_nsg_arrays_refresh_azure():
-    """
-    https://bugzilla.redhat.com/show_bug.cgi?id=1520196
-
-    Polarion:
-        assignee: anikifor
-        casecomponent: Cloud
-        caseimportance: medium
-        initialEstimate: 1/6h
-        testSteps:
-            1. Add Network Security group on Azure with coma separated port ranges
-            `1023,1025` rule inbound/outbound ( ATM this feature is not allowed in
-            East US region of Azure - try West/Central)
-            2. Add such Azure Region into CFME
-            3. Refresh provider
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(2)
-def test_provider_flavors_azure():
-    """
-    Verify that the vm flavors in Azure are of the correct sizes and that
-    the size display in CFME is accurate.  This test checks a regression
-    of https://bugzilla.redhat.com/show_bug.cgi?id=1357086
-    Low priority as it is unlikely to change once set.  Will want to check
-    when azure adds new sizes.  Only need to spot check a few values.
-    For current size values, you can check here:
-    https://azure.microsoft.com/en-us/documentation/articles/virtual-
-    machines-windows-sizes/
-
-    Polarion:
-        assignee: anikifor
-        casecomponent: Cloud
-        caseimportance: low
-        initialEstimate: 1/8h
-        startsin: 5.6
-    """
-    pass
-
-
-@pytest.mark.manual
-@pytest.mark.tier(1)
-def test_market_place_images_azure():
-    """
-    https://bugzilla.redhat.com/show_bug.cgi?id=1491330
-
-    Polarion:
-        assignee: anikifor
-        casecomponent: Cloud
-        caseimportance: medium
-        initialEstimate: 1/6h
-        testSteps:
-            1.Enable market place images
-            2.Verify the list of images
-    """
-    pass
 
 
 @pytest.mark.manual
